requirements.txt
streamlit
xai-sdk
elevenlabs
psycopg[binary]
psycopg2-binary
